Deduplicate same code...
